Timestamp verify and add: move common code to the same script
