+++ website/static/js/pages/timestamp-page.js
@@ -3,6 +3,7 @@
 var $ = require('jquery');
 var nodeApiUrl = window.contextVars.node.urls.api;
 var projectFileList = window.contextVars.project_file_list;
+var timestampCommon = require('./timestamp-common.js');
 
 
 $(document).ready(function () {
@@ -10,11 +11,11 @@ $(document).ready(function () {
     for (var i = 0; i < projectFileList.length; i++) {
         var fileList = projectFileList[i].error_list;
         var providerTr = '<tr><td colspan="5"><b>' + projectFileList[i].provider + '</b></td></tr>';
-        $(providerTr).appendTo($('#tree_timestamp_error_data'));
+        $(providerTr).appendTo($('#timestamp_error_list'));
         for (var j = 0; j < fileList.length; j++) {
             var errorTr =
-                '<tr>' +
-                '<td class="addTimestamp">' +
+                '<tr class="addTimestamp">' +
+                '<td>' +
                 '<input type="checkBox" id="addTimestampCheck" style="width: 15px; height: 15px;" value="' + index + '"/>' +
                 '<td>' + fileList[j].file_path + '</td>' +
                 '<input type="hidden" id="provider" value="' + projectFileList[i].provider + '" />' +
@@ -27,7 +28,7 @@ $(document).ready(function () {
                 '<td>' + fileList[j].operator_date + '</td>' +
                 '<td>' + fileList[j].verify_result_title + '</td>' +
                 '</tr>';
-            $(errorTr).appendTo($('#tree_timestamp_error_data'));
+            $(errorTr).appendTo($('#timestamp_error_list'));
             index++;
         }
     }
@@ -38,76 +39,10 @@ $(function () {
         if ($('#btn-verify').attr('disabled') != undefined || $('#btn-addtimestamp').attr('disabled') != undefined) {
             return false;
         }
-
-        $('#btn-verify').attr('disabled', true);
-        $('#btn-addtimestamp').attr('disabled', true);
-        $('#timestamp_errors_spinner').text('Storage files list gathering ...');
-        var postData = {};
-        var fileCnt = 0;
-        $.ajax({
-            beforeSend: function () {
-                $('#timestamp_errors_spinner').show();
-            },
-            url: 'json/',
-            data: postData,
-            dataType: 'json'
-        }).done(function (projectFileList) {
-            var i, j;
-            projectFileList = projectFileList.provider_list;
-            for (i = 0; i < projectFileList.length; i++) {
-                var fileList = projectFileList[i].provider_file_list;
-                for (j = 0; j < fileList.length; j++) {
-                    fileCnt++;
-                }
-            }
-            var index = 0;
-            var successCnt = 0;
-            for (i = 0; i < projectFileList.length; i++) {
-                fileList = projectFileList[i].provider_file_list;
-                for (j = 0; j < fileList.length; j++) {
-                    var postData = {
-                        'provider': projectFileList[i].provider,
-                        'file_id': fileList[j].file_id,
-                        'file_path': fileList[j].file_path,
-                        'file_name': fileList[j].file_name,
-                        'version': fileList[j].version
-                    };
-                    $.ajax({
-                        url:  nodeApiUrl + 'timestamp/timestamp_error_data/',
-                        data: postData,
-                        dataType: 'json'
-                    }).done(function () {
-                        successCnt++;
-                        $('#timestamp_errors_spinner').text('Verification files : ' + successCnt + ' / ' + fileCnt + ' ...');
-                        if (successCnt == fileCnt) {
-                            $('#timestamp_errors_spinner').text('Verification (100%) and Refreshing...');
-                            window.location.reload();
-                        }
-                    }).fail(function (xhr, status, error) {
-                        $('#btn-verify').removeAttr('disabled');
-                        $('#btn-addtimestamp').removeAttr('disabled');
-                        $('#timestamp_errors_spinner').text('Error : ' + fileList[j].file_path);
-                        Raven.captureMessage('Timestamp Add Error: ' + filePathList[index], {
-                            extra: {
-                                url: url,
-                                status: status,
-                                error: error
-                            }
-                        });
-                    });
-                }
-            }
-        }).fail(function (xhr, textStatus, error) {
-            $('#btn-verify').removeAttr('disabled');
-            $('#btn-addtimestamp').removeAttr('disabled');
-            $('#timestamp_errors_spinner').text('Error : Storage files list gathering Failed');
-            Raven.captureMessage('Timestamp Add Error', {
-                extra: {
-                    url: url,
-                    textStatus: textStatus,
-                    error: error
-                }
-            });
+        timestampCommon.verify({
+            urlVerify: 'json/',
+            urlVerifyData: nodeApiUrl + 'timestamp/timestamp_error_data/',
+            method: 'GET'
         });
     };
 
@@ -115,76 +50,10 @@ $(function () {
         if ($('#btn-verify').attr('disabled') != undefined || $('#btn-addtimestamp').attr('disabled') != undefined) {
             return false;
         }
-
-        var inputCheckBoxs = $('[id=addTimestampCheck]:checked').map(function () {
-            return $(this).val();
+        timestampCommon.add({
+            url: nodeApiUrl + 'timestamp/add_timestamp/',
+            method: 'GET'
         });
-
-        if (inputCheckBoxs.length == 0) {
-            return false;
-        }
-
-        var providerList = $('[id=provider]').map(function () {
-            return $(this).val();
-        });
-
-        var fileIdList = $('[id="file_id"]').map(function () {
-            return $(this).val();
-        });
-
-        var filePathList = $('[id=file_path]').map(function () {
-            return $(this).val();
-        });
-
-        var versionList = $('[id=version]').map(function () {
-            return $(this).val();
-        });
-
-        var fileNameList = $('[id=file_name]').map(function () {
-            return $(this).val();
-        });
-
-        $('#btn-verify').attr('disabled', true);
-        $('#btn-addtimestamp').attr('disabled', true);
-        $('#timestamp_errors_spinner').text('Addtimestamp loading ...');
-        var successCnt = 0;
-        var index;
-        for (var i = 0; i < inputCheckBoxs.length; i++) {
-            index = inputCheckBoxs[i];
-            var postData = {
-                'provider': providerList[index],
-                'file_id': fileIdList[index],
-                'file_path': filePathList[index],
-                'file_name': fileNameList[index],
-                'version': versionList[index]
-            };
-            $.ajax({
-                beforeSend: function () {
-                    $('#timestamp_errors_spinner').show();
-                },
-                url: nodeApiUrl + 'timestamp/add_timestamp/',
-                data: postData,
-                dataType: 'json'
-            }).done(function () {
-                successCnt++;
-                $('#timestamp_errors_spinner').text('Adding Timestamp files : ' + successCnt + ' / ' + inputCheckBoxs.length + ' ...');
-                if (successCnt ==  inputCheckBoxs.length) {
-                    $('#timestamp_errors_spinner').text('Added Timestamp (100%) and Refreshing...');
-                    window.location.reload();
-                }
-            }).fail(function (xhr, textStatus, error) {
-                $('#btn-verify').removeAttr('disabled');
-                $('#btn-addtimestamp').removeAttr('disabled');
-                $('#timestamp_errors_spinner').text('Error : Timestamp Add Failed');
-                Raven.captureMessage('Timestamp Add Error: ' + filePathList[index], {
-                    extra: {
-                        url: url,
-                        textStatus: textStatus,
-                        error: error
-                    }
-                });
-            });
-        }
     };
 
     $('#addTimestampAllCheck').on('change', function () {
